[I-267] Fixed grading
