+++ polymorphia-frontend/app/(logged-in)/course/groups/[courseGroupId]/page.tsx
@@ -1,35 +1,12 @@
 "use client";
 
 import { CourseGroupsProvider } from "@/providers/course-groups";
-import CourseGroups from "@/views/course-groups";
+import CourseGroupsView from "@/views/course-groups";
 import { EquipmentProvider } from "@/providers/equipment";
-import EquipmentModals from "@/components/equipment/modals";
-import GradeModal from "@/components/speed-dial/modals/grade";
-import useCourseGroupsContext from "@/hooks/contexts/useCourseGroupsContext";
 import { TargetProvider } from "@/providers/target";
 import useCourseGroupTargets from "@/hooks/course/course-group/useCourseGroupTargets";
-import useTargetContext from "@/hooks/contexts/useTargetContext";
 
-function CourseGroupViewContent() {
-  const { gradableEventId, setGradableEventId } = useCourseGroupsContext();
-  const { targetId } = useTargetContext();
-
-  return (
-    <>
-      <CourseGroups />
-      <EquipmentModals targetStudentIdOverride={targetId} />
-      {gradableEventId && (
-        <GradeModal
-          gradableEventIdProp={gradableEventId}
-          onClosedAction={() => setGradableEventId(null)}
-          targetStudentIdOverride={targetId}
-        />
-      )}
-    </>
-  );
-}
-
-export default function CourseGroupView() {
+export default function CourseGroups() {
   return (
     <TargetProvider
       useTargets={useCourseGroupTargets}
@@ -41,7 +18,7 @@ export default function CourseGroupView() {
     >
       <EquipmentProvider>
         <CourseGroupsProvider>
-          <CourseGroupViewContent />
+          <CourseGroupsView />
         </CourseGroupsProvider>
       </EquipmentProvider>
     </TargetProvider>
