+++ polymorphia-frontend/app/(logged-in)/equipment/page.tsx
@@ -1,10 +1,10 @@
 import { EquipmentProvider } from "@/providers/equipment";
-import EquipmentView from "@/views/equipment";
+import EquipmentContent from "@/views/equipment";
 
 export default function Equipment() {
   return (
     <EquipmentProvider>
-      <EquipmentView />
+      <EquipmentContent />
     </EquipmentProvider>
   );
 }
